move util test function to config_test.go
