+++ cmd/config/delete_test.go
@@ -2,33 +2,13 @@ package config
 
 import (
 	"bytes"
-	"fmt"
-	"io"
-	"math/rand"
 	"os"
 	"testing"
-	"time"
 
 	"github.com/screwdriver-cd/sd-local/config"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func createRandNameConfig(conf io.Reader) (string, error) {
-	rand.Seed(time.Now().UnixNano())
-	cnfPath := fmt.Sprintf("%vconfig", rand.Int())
-	f, err := os.Create(cnfPath)
-	if err != nil {
-		return "", err
-	}
-	defer f.Close()
-	_, err = io.Copy(f, conf)
-	if err != nil {
-		os.Remove(cnfPath)
-		return "", err
-	}
-	return cnfPath, nil
-}
 
 func TestConfigDeleteCmd(t *testing.T) {
 	f, err := os.Open("./testdata/config")
